requirements.txt
google_api_python_client==2.81.0
google_auth==2.16.2
google_api_core==2.11.0
httplib2==0.21.0
googleapis_common_protos==1.58.0
protobuf==4.22.1
pyasn1_modules==0.2.8
six==1.16.0
cachetools==5.3.0
rsa==4.9
pyparsing==3.0.9
charset_normalizer==3.1.0
idna==3.4
certifi==2022.12.7
urllib3==1.26.15
oauth2client==4.1.3
google_auth_oauthlib==1.0.0
requests_oauthlib==1.3.1
oauthlib==3.2.2
gspread==5.7.2
Django==4.0.6
psycopg2_binary==2.9.5
django_tables2==2.5.3
python-dateutil==1.5
gunicorn==20.1.0

